+++ src/core/quantum/types.py
@@ -20,16 +20,16 @@ class QuantumState:
 
     def __post_init__(self):
         """Ensure state normalization and proper tensor types."""
-        # Convert to complex64 if not already
+        # Convert to complex128 if not already
         if not torch.is_complex(self.amplitudes):
-            self.amplitudes = self.amplitudes.to(torch.complex64)
-        elif self.amplitudes.dtype != torch.complex64:
-            self.amplitudes = self.amplitudes.to(torch.complex64)
+            self.amplitudes = self.amplitudes.to(torch.complex128)
+        elif self.amplitudes.dtype != torch.complex128:
+            self.amplitudes = self.amplitudes.to(torch.complex128)
             
         if not torch.is_complex(self.phase):
-            self.phase = self.phase.to(torch.complex64)
-        elif self.phase.dtype != torch.complex64:
-            self.phase = self.phase.to(torch.complex64)
+            self.phase = self.phase.to(torch.complex128)
+        elif self.phase.dtype != torch.complex128:
+            self.phase = self.phase.to(torch.complex128)
 
         # Normalize state vector
         if len(self.amplitudes.shape) == 1:
@@ -43,9 +43,9 @@ class QuantumState:
 
         # Store original norm if not provided
         if self.original_norm is None:
-            self.original_norm = norm.to(torch.float32)
+            self.original_norm = norm.to(torch.float64)
         else:
-            self.original_norm = self.original_norm.to(torch.float32)
+            self.original_norm = self.original_norm.to(torch.float64)
 
     @property
     def shape(self) -> Tuple[int, ...]:
@@ -60,8 +60,8 @@ class QuantumState:
     def norm(self) -> torch.Tensor:
         """Compute the norm of the quantum state."""
         if len(self.amplitudes.shape) == 1:
-            return torch.sqrt(torch.sum(torch.abs(self.amplitudes) ** 2)).to(torch.float32)
-        return torch.sqrt(torch.sum(torch.abs(self.amplitudes) ** 2, dim=1)).to(torch.float32)
+            return torch.sqrt(torch.sum(torch.abs(self.amplitudes) ** 2)).to(torch.float64)
+        return torch.sqrt(torch.sum(torch.abs(self.amplitudes) ** 2, dim=1)).to(torch.float64)
 
     def density_matrix(self) -> torch.Tensor:
         """Compute the density matrix representation of the state."""
@@ -76,7 +76,7 @@ class QuantumState:
             seq_len = self.amplitudes.shape[1]
             state_dim = self.amplitudes.shape[2]
             rho = torch.zeros((batch_size, seq_len, state_dim, state_dim), 
-                             dtype=torch.complex64, 
+                             dtype=torch.complex128, 
                              device=self.amplitudes.device)
             for i in range(batch_size):
                 for j in range(seq_len):
@@ -86,7 +86,7 @@ class QuantumState:
             # Regular batch case
             state_dim = self.amplitudes.shape[1]
             rho = torch.zeros((batch_size, state_dim, state_dim), 
-                             dtype=torch.complex64, 
+                             dtype=torch.complex128, 
                              device=self.amplitudes.device)
             for i in range(batch_size):
                 rho[i] = torch.outer(self.amplitudes[i], self.amplitudes[i].conj())
